fix errors, removed csv file, corrected readme
