+++ protocols/Imperial_College/drug_mix_transfer.ot1.py
@@ -38,32 +38,34 @@ def get_csv_content(csv_string):
     csv_list = csv_to_list(csv_string)
     headers = csv_list[0:27]
 
-    for i in range(1, int(len(csv_list)/27)):
-        for j in range(0, len(headers)):
-            if headers[j] in content:
+    for column_num in range(1, int(len(csv_list) / 27)):
+        for header_num in range(0, len(headers)):
+            if headers[header_num] in content:
                 pass
             else:
-                content[headers[j]] = []
-            content[headers[j]].append(csv_list[i*27+j])
+                content[headers[header_num]] = []
+            content[headers[header_num]].append(
+                csv_list[column_num * 27 + header_num])
 
     new_dict = {}
-    for i in range(3, 27):
-        for j in range(0, len(content[headers[i]])):
-            if content['Plate'][j] == '1':
-                content['Plate'][j] = comp_plate_1
-            elif content['Plate'][j] == '2':
-                content['Plate'][j] = comp_plate_2
-            elif content['Plate'][j] == '3':
-                content['Plate'][j] = comp_plate_3
+    for well_num in range(3, 27):
+        for compound in range(0, len(content[headers[well_num]])):
+            if content['Plate'][compound] == '1':
+                content['Plate'][compound] = comp_plate_1
+            elif content['Plate'][compound] == '2':
+                content['Plate'][compound] = comp_plate_2
+            elif content['Plate'][compound] == '3':
+                content['Plate'][compound] = comp_plate_3
 
-        new_dict[headers[i]] = (list(zip(content[headers[i]], content['Plate'],
-                                content['Well'])))
+        new_dict[headers[well_num]] = (list(zip(
+            content[headers[well_num]], content['Plate'], content['Well'])))
+
         del_list = []
-        for k in range(0, len(new_dict[headers[i]])):
-            if not new_dict[headers[i]][k][0]:
-                del_list.append(k)
-        for m in del_list:
-            del new_dict[headers[i]][m]
+        for empty_volume in range(0, len(new_dict[headers[well_num]])):
+            if not new_dict[headers[well_num]][empty_volume][0]:
+                del_list.append(empty_volume)
+        for empty_compound in del_list:
+            del new_dict[headers[well_num]][empty_compound]
 
     return headers, new_dict
 
@@ -79,14 +81,17 @@ comp_3,1,C1,1,2,3,4,5,6,7,8,9,10,1,2,3,4,5,6,7,8,9,10,1,2,3,4
 def run_custom_protocol(volumes_csv: FileInput=example_volumes_csv):
 
     headers, transfer_dict = get_csv_content(volumes_csv)
+    print(headers)
+    print(transfer_dict)
 
-    for i in headers[3:]:
+    for well in headers[3:]:
         trasnfer_vol = 0
-        for j in transfer_dict[i][:-1]:
-            p10.transfer(int(j[0]), j[1][2], mix_plate[i])
-            trasnfer_vol += int(j[0])
-        else:
-            p10.transfer(int(j[0]), j[1][2], mix_plate[i],
-                         mix_after=(2, int(j[0])))
-            trasnfer_vol += int(j[0])
-            p10.transfer(trasnfer_vol, mix_plate[i], target_plate[i])
+        count = 1
+        for each_compound in transfer_dict[well]:
+            volume = int(each_compound[0])
+            source = each_compound[1][each_compound[2]]
+            p10.transfer(volume, source, mix_plate[well])
+            trasnfer_vol += volume
+            if count == len(transfer_dict[well]):
+                p10.transfer(trasnfer_vol, mix_plate[well], target_plate[well])
+            count += 1
